pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.run]
branch = true
source =["jsonschema_spec"]

[tool.coverage.xml]
output = "reports/coverage.xml"

[tool.mypy]
files = "jsonschema_spec"
strict = true

[[tool.mypy.overrides]]
module = "jsonschema.validators"
ignore_missing_imports = true

[tool.poetry]
name = "jsonschema-spec"
version = "0.1.0"
description = "JSONSchema Spec with object-oriented paths"
authors = ["Artur Maciag <maciag.artur@gmail.com>"]
license = "Apache License, Version 2.0"
readme = "README.rst"
repository = "https://github.com/p1c2u/jsonschema-spec"
keywords = ["jsonschema", "swagger", "spec"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Topic :: Software Development :: Libraries"
]

[tool.poetry.dependencies]
jsonschema = "^4.0.0"
pathable = "^0.4.1"
python = "^3.7.0"
PyYAML = ">=5.1"
typing-extensions = "^4.3.0"

[tool.poetry.extras]
dev = ["pre-commit"]

[tool.poetry.dev-dependencies]
pre-commit =  {version = "*", optional = true}
pytest = "^6.2.5"
pytest-flake8 = "=1.0.7"
pytest-cov = "^3.0.0"
isort = "^5.0.0"
black = "^22.0.0"
flynt = "0.76"
mypy = "^0.971"
types-PyYAML = "^6.0.11"
types-requests = "^2.28.9"

[tool.pytest.ini_options]
addopts = """
--capture=no
--verbose
--showlocals
--junitxml=reports/junit.xml
--cov=jsonschema_spec
--cov-report=term-missing
--cov-report=xml
"""

[tool.black]
line-length = 79

[tool.isort]
profile = "black"
line_length = 79
force_single_line = true
